refactored elasticsearch to be a controller
