+++ app/src/main/java/cmput301f17t01/bronzify/controllers/LoginController.java
@@ -2,7 +2,6 @@ package cmput301f17t01.bronzify.controllers;
 
 import cmput301f17t01.bronzify.models.AppLocale;
 import cmput301f17t01.bronzify.models.Controller;
-import cmput301f17t01.bronzify.models.ElasticSearch;
 import cmput301f17t01.bronzify.models.User;
 
 /**
