+++ app/src/main/java/cmput301f17t01/bronzify/controllers/ProfileController.java
@@ -1,12 +1,9 @@
 package cmput301f17t01.bronzify.controllers;
 
-import android.app.AlertDialog;
 import android.util.Log;
 
-import cmput301f17t01.bronzify.activities.MyProfileActivity;
 import cmput301f17t01.bronzify.models.AppLocale;
 import cmput301f17t01.bronzify.models.Controller;
-import cmput301f17t01.bronzify.models.ElasticSearch;
 
 /**
  * Created by kdehaan on 12/11/17.
